pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppp"
version = "0.1.0"
description = "Public Participation Probe (PPP) - Policy-governed autonomous agent harness"
authors = [{name = "Anthropic", email = "ppp@anthropic.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ppp = "src.ppp.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src/ppp --cov-report=term-missing"
